Backend : LTS Enhancend Filtering computations for driver location updates
