+++ dhall_config/location_tracking_service.dhall
@@ -52,6 +52,7 @@ in {
     bulk_location_callback_url = "http://127.0.0.1:8016/internal/bulkLocUpdate",
     auth_token_expiry = 86400,
     min_location_accuracy = 50.0,
+    driver_location_accuracy_buffer = 25.0,
     redis_expiry = 86400,
     last_location_timstamp_expiry = 86400,
     location_update_limit = 6000000000,
@@ -60,5 +61,4 @@ in {
     batch_size = 100,
     bucket_size = 30,
     nearby_bucket_threshold = 4,
-    driver_location_accuracy_buffer = 50.0,
 }
